React : DAY 07 수업 내용

+++ 07. React/DAY06/board-file/server/src/main/java/com/aloha/server/service/FileServiceImpl.java
@@ -1,10 +1,10 @@
 package com.aloha.server.service;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import java.io.FileInputStream;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -20,11 +20,9 @@ import jakarta.servlet.ServletOutputStream;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 
-
 /**
  * FileServiceImpl
  */
-
 @Slf4j
 @Service
 public class FileServiceImpl implements FileService {
@@ -57,7 +55,23 @@ public class FileServiceImpl implements FileService {
 
     @Override
     public int delete(int no) throws Exception {
-        return fileMapper.delete(no);
+        // 1️⃣ 파일 정보 조회
+        Files file = fileMapper.select(no);
+        // 2️⃣ 파일 경로로 파일 객체 접근
+        String filePath = file.getFilePath();
+        File deleteFiie = new File(filePath);
+        // 3️⃣ 파일시스템의 파일 삭제
+        // - 파일 존재여부 확인
+        if( !deleteFiie.exists()) return 0;
+        // 파일 삭제 
+        boolean deleted = deleteFiie.delete();
+        // 4️⃣ DB의 파일 데이터 삭제
+        int result = 0;
+        if( deleted ) {
+            result = fileMapper.delete(no);
+            return result;
+        }
+        return result;
     }
 
     @Override
@@ -134,17 +148,17 @@ public class FileServiceImpl implements FileService {
             response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
             return 0;
         }
-        String filePath = file.getFilePath();
-        String fileName = file.getFileName();
+        String filePath = file.getFilePath();   //파일 경로
+        String fileName = file.getFileName();   //파일 이름
 
-        // 파일 다운로드를 위한 헤더 세팅
-        // - Content-Type : application/octect-stream
+        // 파일 다운로드를 위한 🎫 헤더 세팅
+        // Content-Type : application/octect-stream
         // - Content-Disposition : attachment, filename="파일명.확장자"
         response.setContentType(MediaType.APPLICATION_OCTET_STREAM_VALUE);
-        response.setHeader("Content-Disposition",
-                            "attachment; filename=\"" + fileName + "\"");
+        response.setHeader("Content-Disposition", 
+                            "attatchment; filename=\"" + fileName + "\"");
 
-        // 파일 다운로드
+        // 📄⬇ 파일 다운로드
         // - 파일 입력
         File downloadFile = new File(filePath);
         FileInputStream fis = new FileInputStream(downloadFile);
@@ -154,15 +168,34 @@ public class FileServiceImpl implements FileService {
 
         // - 다운로드
         FileCopyUtils.copy(fis, sos);
-        // byte[] buffer = new byte[1024];
+
+        // byte[] buffer = new byte[1024];             // 1024bytes : 1KB
         // int data;
-        // while( (data = fis.read(buffer)) != -1 ) {
-        //     sos.write(buffer, 0, data);
+        // while ((data = fis.read(buffer)) != 1) {    // 1KB 입력
+        //     sos.write(buffer, 0, data);         // 1KB 출력 (전송)
         // }
+
         fis.close();
         sos.close();
+
         return 1;
     }
 
+    @Override
+    public int deleteFiles(String no) throws Exception {
+        String[] noList = no.split(",");
+        log.info("sdfsdf");
+        int result = 0;
+        for (String deleteNo : noList) {
+            int fileNo = Integer.parseInt(deleteNo.trim());
+            result += delete(fileNo);
+        }
+        return result;
+    }
+
+    @Override
+    public int deleteByParent(Files file) throws Exception {
+        return fileMapper.deleteByParent(file);
+    }
 
 }
